Install vim full version

+++ setup.sh
@@ -1,9 +1,5 @@
 #!/bin/sh
 
-# If you have already installed the customized vim environment,
-# this setup is interrupted here.
-[ `which vim` ] && [ `which git` ] && [ -d ~/.vim/bundle/neobundle.vim ] && exit 0
-
 # Determine package manager
 INSTALLER=""
 if [ `uname` = "Linux" ]; then
@@ -16,6 +12,20 @@ fi
 [ -n "$INSTALLER" ] || exit 1
 # Install Vim and Git
 sudo $INSTALLER install vim git || exit 1
+# Install Vim full version
+if [ ! `which vim.nox` ]; then
+    echo -n "Do you want to install vim full version? [Y/n] : "
+    read answer
+    case "$answer" in
+	N | n | No | no)
+	    echo "Continue..." ;;
+	*)
+	    sudo $INSTALLER install vim-nox ;;
+    esac
+fi
+# If you have already installed the customized vim environment,
+# this setup is interrupted here.
+[ `which vim` ] && [ `which git` ] && [ -d ~/.vim/bundle/neobundle.vim ] && exit 0
 
 # Copy customized .vimrc to your home directory
 cp -i .vimrc ~ || exit 1
